Changed from int to INT8

+++ main.c
@@ -10,31 +10,31 @@
 void playerMovement();
 void drawSprites();
 void loadSprites();
-int rectCollision(int x1, int y1, int w1, int h1, int x2, int y2, int w2, int h2);
-int abs(int x);
-void log(char* m, int data);
+UINT8 rectCollision(INT8 x1, INT8 y1, INT8 w1, INT8 h1, INT8 x2, INT8 y2, INT8 w2, INT8 h2);
+UINT8 abs(INT8 x);
+void log(char* m, UINT8 data);
 
 // Set the base location of the tile map
 unsigned char memoryCounter = 0x1A;
 
 // Set the fps
-int FPS = 16; //60 FPS
-// int FPS = 32; // 30 FPS
-//int FPS = 64; // 15 FPS
+UINT8 FPS = 16; //60 FPS
+// UINT8 FPS = 32; // 30 FPS
+//UINT8 FPS = 64; // 15 FPS
 
 // Store the location of each sprite
-int ballLocation = 0;
+UINT8 ballLocation = 0;
 
 // Store the number of sprites
-const int spriteCount = 2;
+const UINT8 spriteCount = 2;
 
 // Create data structure to hold the sprite data
-int spriteLocations[2] = {0, 1};
-int spriteX[2] = {75, 60};
-int spriteY[2] = {75, 60};
+UINT8 spriteLocations[2] = {0, 1};
+UINT8 spriteX[2] = {75, 60};
+UINT8 spriteY[2] = {75, 60};
 
 // Store the speed of the player (allows for a grid based instead of pixel based movement if set to 8)
-int playerSpeed = 1;
+UINT8 playerSpeed = 1;
 
 void main() {
 
@@ -51,8 +51,8 @@ void main() {
         if (rectCollision(spriteX[0],spriteY[0],8,8,spriteX[1],spriteY[1],8,8)) {
 
             // Move the mushroom
-            int xRand = abs((UINT8)rand());
-            int yRand = abs((UINT8)rand());
+            UINT8 xRand = abs((UINT8)rand());
+            UINT8 yRand = abs((UINT8)rand());
             spriteX[1] = 8 + (xRand % 152);
             spriteY[1] = 16 + (yRand % 136);
             log("X Rand:",xRand);
@@ -108,7 +108,7 @@ void playerMovement() {
 void drawSprites() {
 
     // Look at each sprite
-    int i;
+    UINT8 i;
     for (i = 0; i < spriteCount; i++) {
 
         // Render the sprite (move it to the correct location to be rendered)
@@ -136,15 +136,15 @@ void loadSprites() {
 }
 
 // Create a function to check for rect collision
-int rectCollision(int x1, int y1, int w1, int h1, int x2, int y2, int w2, int h2) {
+UINT8 rectCollision(INT8 x1, INT8 y1, INT8 w1, INT8 h1, INT8 x2, INT8 y2, INT8 w2, INT8 h2) {
 
     // Test for a collision between two rects
-    return abs((int)x1-(int)x2) < w1/2 + w2/2 & abs((int)y1-(int)y2) < h1/2 + h2/2;
+    return abs(x1-x2) < w1/2 + w2/2 & abs(y1-y2) < h1/2 + h2/2;
 
 }
 
 // Function to return an absolute value
-int abs(int x) {
+UINT8 abs(INT8 x) {
     if (x == 0) {
         return 0;
     }
@@ -155,7 +155,7 @@ int abs(int x) {
 }
 
 // Create a procedure to output a log message
-void log(char* m, int data) {
+void log(char* m, UINT8 data) {
 
     printf("%s%d\n",m,(UWORD)data);
